pyproject.toml
[project]
name = "vlax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "decouple>=0.0.7",
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-nested-routers>=0.95.0",
    "environ>=1.0",
    "pillow>=11.3.0",
]
